+++ src/Network/Network/ExpressRouteCircuit/Peering/AzureExpressRouteCircuitPeeringConfigBase.cs
@@ -128,7 +128,7 @@ namespace Microsoft.Azure.Commands.Network
         {
             if (this.MicrosoftConfigAdvertisedPublicPrefixes != null && this.MicrosoftConfigAdvertisedPublicPrefixes.Any())
             {
-                if (PeerAddressType == IPv6)
+                if (string.Equals(this.PeerAddressType, "IPv6", StringComparison.OrdinalIgnoreCase))
                 {
                     peering.Ipv6PeeringConfig.MicrosoftPeeringConfig = new PSPeeringConfig();
                     peering.Ipv6PeeringConfig.MicrosoftPeeringConfig.AdvertisedPublicPrefixes = this.MicrosoftConfigAdvertisedPublicPrefixes?.ToList();