+++ src/Network/Network/ExpressRouteCircuit/Peering/SetAzureExpressRouteCircuitPeeringConfigCommand.cs
@@ -65,7 +65,7 @@ namespace Microsoft.Azure.Commands.Network
                 peering.SharedKey = this.SharedKey;
             }
 
-            if(PeerAddressType == IPv6)
+            if (string.Equals(this.PeerAddressType, "IPv6", StringComparison.OrdinalIgnoreCase))
             {
                 this.SetIpv6PeeringParameters(peering);
             }